Update Application to improve architecture
-- Moved the connectors, handlers and authentication to the right package

-- Moved MQTT status function from handler to connector

-- Created SFTP connector and removed SFTP connection logic from content route

-- Moved SFTP upload logic from content route to content handler

TODO: Further test changes, introduce logger to content route
