+++ authentication/passport_handler.js
@@ -1,6 +1,6 @@
 const passport = require('passport')
 const bcrypt = require('bcrypt');
-const user_handler = require('../../data/user_handler');
+const user_handler = require('../user_handler');
 const passportJWT = require('passport-jwt');
 const extractJWT = passportJWT.ExtractJwt;
 const LocalStrategy = require('passport-local').Strategy;
